+++ src/main/java/roomescape/admin/dto/AdminReservationRequest.java
@@ -1,7 +1,7 @@
 package roomescape.admin.dto;
 
 import java.time.LocalDate;
-import roomescape.reservation.dto.ReservationRequest;
+import roomescape.reservation.dto.request.ReservationRequest;
 
 public record AdminReservationRequest(LocalDate date, Long timeId, Long themeId, Long memberId) {
     public ReservationRequest getReservationRequest() {
