+++ src/main/java/roomescape/member/dto/response/LoginCheckResponse.java
@@ -1,4 +1,4 @@
-package roomescape.member.dto;
+package roomescape.member.dto.response;
 
 public record LoginCheckResponse(String name) {
 }
