+++ src/main/java/roomescape/member/dto/response/ReservationMemberResponse.java
@@ -1,4 +1,4 @@
-package roomescape.member.dto;
+package roomescape.member.dto.response;
 
 public record ReservationMemberResponse(String name) {
 }
