+++ src/main/java/roomescape/member/presentation/MemberController.java
@@ -10,9 +10,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.member.dto.MemberResponse;
-import roomescape.member.dto.SignupRequest;
-import roomescape.member.dto.SignupResponse;
+import roomescape.member.dto.response.MemberResponse;
+import roomescape.member.dto.request.SignupRequest;
+import roomescape.member.dto.response.SignupResponse;
 import roomescape.member.service.MemberService;
 
 @RestController
