+++ src/main/java/roomescape/reservation/dto/response/ReservationResponse.java
@@ -1,10 +1,10 @@
-package roomescape.reservation.dto;
+package roomescape.reservation.dto.response;
 
 import java.time.LocalDate;
-import roomescape.member.dto.ReservationMemberResponse;
+import roomescape.member.dto.response.ReservationMemberResponse;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservationTime.dto.ReservationTimeResponse;
-import roomescape.theme.dto.ThemeResponse;
+import roomescape.reservationTime.dto.response.ReservationTimeResponse;
+import roomescape.theme.dto.response.ThemeResponse;
 
 public record ReservationResponse(Long id, ReservationMemberResponse member, LocalDate date,
                                   ReservationTimeResponse time,
