+++ src/main/java/roomescape/reservationTime/service/ReservationTimeService.java
@@ -6,10 +6,10 @@ import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationRepository;
 import roomescape.reservationTime.domain.ReservationTime;
 import roomescape.reservationTime.domain.ReservationTimeRepository;
-import roomescape.reservationTime.dto.ReservationTimeRequest;
-import roomescape.reservationTime.dto.ReservationTimeResponse;
-import roomescape.reservationTime.dto.TimeConditionRequest;
-import roomescape.reservationTime.dto.TimeConditionResponse;
+import roomescape.reservationTime.dto.request.ReservationTimeRequest;
+import roomescape.reservationTime.dto.response.ReservationTimeResponse;
+import roomescape.reservationTime.dto.request.TimeConditionRequest;
+import roomescape.reservationTime.dto.response.TimeConditionResponse;
 
 @Service
 public class ReservationTimeService {
