+++ src/main/java/roomescape/theme/dto/request/ThemeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.theme.dto;
+package roomescape.theme.dto.request;
 
 public record ThemeRequest(String name, String description, String thumbnail) {
     public ThemeRequest {
