+++ src/main/java/roomescape/theme/dto/response/PopularThemeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.theme.dto;
+package roomescape.theme.dto.response;
 
 public record PopularThemeResponse(String name, String thumbnail, String description) {
 }
