+++ src/test/java/roomescape/member/dto/SignupRequestTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+import roomescape.member.dto.request.SignupRequest;
 
 class SignupRequestTest {
