+++ src/test/java/roomescape/member/service/LoginServiceTest.java
@@ -15,8 +15,8 @@ import roomescape.common.util.SystemDateTime;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberRepository;
 import roomescape.member.domain.Role;
-import roomescape.member.dto.LoginMember;
-import roomescape.member.dto.LoginRequest;
+import roomescape.member.dto.request.LoginMember;
+import roomescape.member.dto.request.LoginRequest;
 
 class LoginServiceTest {
