+++ src/test/java/roomescape/theme/service/ThemeServiceTest.java
@@ -19,7 +19,7 @@ import roomescape.reservation.service.FakeReservationRepository;
 import roomescape.reservation.service.FakeThemeRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.ThemeRepository;
-import roomescape.theme.dto.PopularThemeResponse;
+import roomescape.theme.dto.response.PopularThemeResponse;
 
 class ThemeServiceTest {
